# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenMapX API",
    "description": "Generated from the API's own route table and the integration route registry. Do not edit by hand — run `pnpm openapi:generate`.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/integrations/overlay-satellite/tiles/{layerId}/{date}/{z}/{y}/{x}": {
      "get": {
        "operationId": "getApiIntegrationsOverlaySatelliteTilesLayerIdDateZYX",
        "parameters": [
          {
            "name": "layerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /tiles/:layerId/:date/:z/:y/:x (overlay-satellite)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "overlay-satellite",
        "x-openmapx-source": "integrations/overlay-satellite/index.ts"
      }
    },
...
    "/api/integrations/overlay-weather/radar/tile/{z}/{x}/{y}": {
      "get": {
        "operationId": "getApiIntegrationsOverlayWeatherRadarTileZXY",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /radar/tile/:z/:x/:y (overlay-weather)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "overlay-weather",
        "x-openmapx-source": "integrations/overlay-weather/index.ts"
      }
    },
...
    "/api/integrations/street-level-imagery-mapillary/tiles/{z}/{x}/{y}": {
      "get": {
        "operationId": "getApiIntegrationsStreetLevelImageryMapillaryTilesZXY",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /tiles/:z/:x/:y (street-level-imagery-mapillary)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "street-level-imagery-mapillary",
        "x-openmapx-source": "integrations/street-level-imagery-mapillary/index.ts"
      }
    },
...
    "/api/integrations/street-level-imagery-panoramax/tiles/{z}/{x}/{y}": {
      "get": {
        "operationId": "getApiIntegrationsStreetLevelImageryPanoramaxTilesZXY",
        "parameters": [
          {
            "name": "z",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /tiles/:z/:x/:y (street-level-imagery-panoramax)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "street-level-imagery-panoramax",
        "x-openmapx-source": "integrations/street-level-imagery-panoramax/index.ts"
      }
    },
...
    "/api/integrations/transit/routes/{id}": {
      "get": {
        "operationId": "getApiIntegrationsTransitRoutesId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Integration response"
          }
        },
        "summary": "GET /routes/:id (transit)",
        "tags": [
          "integrations"
        ],
        "x-openmapx-auth": "public",
        "x-openmapx-integration": "transit",
        "x-openmapx-source": "integrations/transit/index.ts"
      }
    },
...
    "/api/maptiler/{wildcard}": {
      "get": {
        "operationId": "getApiMaptilerWildcard",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "*",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "maptiler"
        ],
        "x-openmapx-auth": "public"
      }
    },
...
    "/api/osm/contributions/{type}/{id}": {
      "get": {
        "operationId": "getApiOsmContributionsTypeId",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "type",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        },
        "tags": [
          "osm"
        ],
        "x-openmapx-auth": "session"
      }
    },
...
  }
}